+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.3+3]
+
+### Added
+
+- Added custom file `extension` for recursive Data Class Generator.
+
+### Fixed
+
+- Fixed minor equality bug in I18N Generator.
+
+### Changed
+
+- Updated dependencies.
+
 ## [1.0.3+2]
 
 ### Fixed
@@ -61,7 +75,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release including Assets, I18N, Data Class and Icon Fonts generators.
 
-[unreleased]: https://github.com/Jlgtri/generators_lite/compare/v1.0.3+1...HEAD
+[unreleased]: https://github.com/Jlgtri/generators_lite/compare/v1.0.3+3...HEAD
+[1.0.3+3]: https://github.com/Jlgtri/generators_lite/compare/v1.0.3+2..v1.0.3+3
+[1.0.3+2]: https://github.com/Jlgtri/generators_lite/compare/v1.0.3+1..v1.0.3+2
 [1.0.3+1]: https://github.com/Jlgtri/generators_lite/compare/v1.0.3..v1.0.3+1
 [1.0.3]: https://github.com/Jlgtri/generators_lite/compare/v1.0.2..v1.0.3
 [1.0.2]: https://github.com/Jlgtri/generators_lite/compare/v1.0.1+3..v1.0.2